requirements.txt
filelock==3.14.0
fsspec==2024.5.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
sympy==1.12
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.12.0
